requirements.txt
datasets==4.0.0
fastapi==0.116.1
gdown==5.2.0
google_api_python_client==2.177.0
google_auth_oauthlib==1.2.2
ipython==8.12.3
langchain==0.3.27
langchain_community==0.3.27
langchain_core==0.3.74
langchain_huggingface==0.3.1
langchain_openai==0.3.30
langchain_text_splitters==0.3.9
Markdown==3.8.2
openai==1.100.2
pandas==2.3.1
peft==0.16.0
protobuf==6.32.0
pydantic==2.11.7
python-dotenv==1.1.1
scikit_learn==1.7.1
streamlit==1.48.0
tqdm==4.67.1
